Hover: default variable type to Any.
It likely makes sense to only show Any as the
type when there is no annotation, rather than
attempting to show an inferred type.
This is because Any is the default type assigned
to the variable, and so assignments to that
variable are checked against the Any type rather
than the assigned value's effective type.

Signed-off-by: Adam Blanchet <adamblanchet@free.fr>